requirements.txt
autopep8==1.6.0
boto3==1.23.10
botocore==1.26.10
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
jmespath==1.0.0
lxml==4.8.0
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
urllib3==1.26.9
